docker

+++ src/types/types.controller.ts
@@ -1,4 +1,13 @@
-import { Controller, Get, Post, Body, Param, Delete } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Param,
+  Delete,
+  Res,
+  HttpStatus,
+} from '@nestjs/common';
 import { TypesService } from './types.service';
 import { CreateTypeDto } from './dto/create-type.dto';
 
@@ -7,22 +16,43 @@ export class TypesController {
   constructor(private readonly typesService: TypesService) {}
 
   @Post()
-  create(@Body() createTypeDto: CreateTypeDto) {
-    return this.typesService.create(createTypeDto);
+  async create(@Res() response, @Body() createTypeDto: CreateTypeDto) {
+    const type = await this.typesService.findOne(createTypeDto.label);
+
+    if (type) {
+      return response.status(HttpStatus.BAD_REQUEST).json({
+        message: `${createTypeDto.label} already exists`,
+      });
+    }
+
+    const newType = await this.typesService.create(createTypeDto);
+    return response.status(HttpStatus.CREATED).json(newType);
   }
 
   @Get()
-  findAll() {
-    return this.typesService.findAll();
+  async findAll(@Res() response) {
+    return response
+      .status(HttpStatus.OK)
+      .json(await this.typesService.findAll());
   }
 
   @Get(':label')
-  findOne(@Param('label') label: string) {
-    return this.typesService.findOne(label);
+  async findOne(@Res() response, @Param('label') label: string) {
+    const foundType = await this.typesService.findOne(label);
+    if (!foundType) {
+      return response.status(HttpStatus.NOT_FOUND).json();
+    }
+    return response.status(HttpStatus.OK).json(foundType);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.typesService.remove(id);
+  async remove(@Res() response, @Param('id') id: string) {
+    const foundType = await this.typesService.findOneById(id);
+    if (!foundType) {
+      return response.status(HttpStatus.NOT_FOUND).json();
+    }
+
+    await this.typesService.remove(id);
+    return response.status(HttpStatus.NO_CONTENT).json();
   }
 }
